+++ src/main/scala/edu/rit/cs/mmior/jsonoid/generator/ArrayGenerator.scala
@@ -13,10 +13,15 @@ import edu.rit.cs.mmior.jsonoid.discovery.schemas.{
 import org.json4s._
 
 object ArrayGenerator extends Generator[ArraySchema, JArray] {
-  def generate(schema: ArraySchema, depth: Int): JArray =
-    generate(schema, depth, None)
+  def generate(schema: ArraySchema, depth: Int, useExamples: Boolean): JArray =
+    generate(schema, depth, useExamples, None)
 
-  def generate(schema: ArraySchema, depth: Int, items: Option[Int]): JArray = {
+  def generate(
+      schema: ArraySchema,
+      depth: Int,
+      useExamples: Boolean,
+      items: Option[Int]
+  ): JArray = {
     val itemType = schema.properties.get[ItemTypeProperty].itemType
 
     itemType match {
@@ -55,17 +60,25 @@ object ArrayGenerator extends Generator[ArraySchema, JArray] {
         JArray(if (isUnique) {
           Stream
             .from(1)
-            .map(_ => Generator.generateFromSchema(itemSchema, depth + 1))
+            .map(_ =>
+              Generator.generateFromSchema(itemSchema, depth + 1, useExamples)
+            )
             .distinct
             .take(numItems)
             .toList
         } else {
           (1 to numItems)
-            .map(_ => Generator.generateFromSchema(itemSchema, depth + 1))
+            .map(_ =>
+              Generator.generateFromSchema(itemSchema, depth + 1, useExamples)
+            )
             .toList
         })
       case Right(schemas) =>
-        JArray(schemas.map(s => Generator.generateFromSchema(s, depth + 1)))
+        JArray(
+          schemas.map(s =>
+            Generator.generateFromSchema(s, depth + 1, useExamples)
+          )
+        )
     }
   }
 }