+++ src/main/scala/edu/rit/cs/mmior/jsonoid/generator/EnumGenerator.scala
@@ -8,9 +8,9 @@ import edu.rit.cs.mmior.jsonoid.discovery.schemas.{
 import org.json4s._
 
 object EnumGenerator extends Generator[EnumSchema, JValue] {
-  def generate(schema: EnumSchema, depth: Int): JValue = {
+  def generate(schema: EnumSchema, depth: Int, useExamples: Boolean): JValue = {
+    // Note: useExamples is ignored here since we're just picking from the enum
     val values = schema.properties.get[EnumValuesProperty].values
-
     values.iterator.drop(util.Random.nextInt(values.size)).next
   }
 }