+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `patternProperties`
 - Partial `ReferenceSchema` support
 - Limit maximum depth of generated objects
+- Setting to generate values only from examples
 
 ### Changed
 - Updated jsonoid-discovery dependency to v0.7.3